+++ pkg/providers/cloudaccess/cloudaccess_errors.go
@@ -5,4 +5,6 @@ import "errors"
 var (
 	// ErrCloudAccessKey is returned when operation on cloud access key fails
 	ErrCloudAccessKey = errors.New("cloud access key")
+	// ErrCloudAccessKeyProperties is returned when operation on cloud access key properties fails
+	ErrCloudAccessKeyProperties = errors.New("cloud access key properties")
 )