+++ pkg/providers/cloudaccess/testdata/TestDataKeyProperties/invalid.tf
@@ -1,0 +1,5 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_cloudaccess_key_properties" "test" {}